+++ pages/styleguide.js
@@ -1,17 +1,17 @@
 import React from 'react'
-import Page from './Page'
-import InnerHeader from './InnerHeader'
-import Box from './Box'
-import Container from './Container'
-import Heading from './Heading'
-import Text from './Text'
-import Caps from './Caps'
-import Button from './Button'
-import NavLink from './NavLink'
-import Image from './Image'
-import BackgroundImage from './BackgroundImage'
-import Avatar from './Avatar'
-import Logos from './Logos'
+import Page from '../src/Page'
+import InnerHeader from '../src/InnerHeader'
+import Box from '../src/Box'
+import Container from '../src/Container'
+import Heading from '../src/Heading'
+import Text from '../src/Text'
+import Caps from '../src/Caps'
+import Button from '../src/Button'
+import NavLink from '../src/NavLink'
+import Image from '../src/Image'
+import BackgroundImage from '../src/BackgroundImage'
+import Avatar from '../src/Avatar'
+import Logos from '../src/Logos'
 
 const Section = ({ title, children }) => (
   <Box py={4}>
